requirements.txt
django>=6.0
gunicorn>=21.2
whitenoise>=6.7
djangorestframework>=3.17
djangorestframework-simplejwt>=5.5
django-cors-headers>=4.9
django-ratelimit>=4.1
psycopg2-binary>=2.9
redis>=5.0
celery>=5.3
pdfplumber>=0.11
pikepdf>=8.0
boto3>=1.34
requests>=2.31
google-generativeai>=0.8
python-dotenv>=1.0
pytz>=2024.1
python-dateutil>=2.9
